requirements.txt
beautifulsoup4
requests
requests>=2.28.0
pandas>=1.5.0
python-slugify>=8.0.0
python-dotenv>=0.21.0
